Adjustments to the Comparison code for external algorithm aiming to publish.
