+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/algorithm/external/ExternalAlgorithmBNTPc.java
@@ -48,14 +48,10 @@ import java.util.List;
  *
  * @author jdramsey
  */
-public class ExternalAlgorithmBNTPc implements ExternalAlgorithm {
+public class ExternalAlgorithmBNTPc extends ExternalAlgorithm {
     static final long serialVersionUID = 23L;
     private final String extDir;
     private String shortDescription = null;
-    private String path;
-    private List<String> usedParameters = new ArrayList<>();
-    private Simulation simulation;
-    private int simIndex = -1;
 
     public ExternalAlgorithmBNTPc(String extDir) {
         this.extDir = extDir;
@@ -67,7 +63,6 @@ public class ExternalAlgorithmBNTPc implements ExternalAlgorithm {
         this.shortDescription = shortDecription;
     }
 
-    @Override
     /**
      * Reads in the relevant graph from the file (see above) and returns it.
      */
@@ -104,13 +99,11 @@ public class ExternalAlgorithmBNTPc implements ExternalAlgorithm {
         }
     }
 
-    @Override
     /**
      * Returns the pattern of the supplied DAG.
      */
     public Graph getComparisonGraph(Graph graph) {
         return new EdgeListGraph(graph);
-//        return SearchGraphUtils.patternForDag(new EdgeListGraph(graph));
     }
 
     public String getDescription() {
@@ -122,34 +115,8 @@ public class ExternalAlgorithmBNTPc implements ExternalAlgorithm {
     }
 
     @Override
-    public List<String> getParameters() {
-        return usedParameters;
-    }
-
-    public int getNumDataModels() {
-        return simulation.getNumDataModels();
-    }
-
-    @Override
     public DataType getDataType() {
         return DataType.Continuous;
-    }
-
-    public void setSimulation(Simulation simulation) {
-        this.simulation = simulation;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public void setSimIndex(int simIndex) {
-        this.simIndex = simIndex;
-    }
-
-    @Override
-    public Simulation getSimulation() {
-        return simulation;
     }
 
     @Override
@@ -168,8 +135,6 @@ public class ExternalAlgorithmBNTPc implements ExternalAlgorithm {
         }
 
         File file = new File(path, "/elapsed/" + extDir + "/" + (simIndex + 1) + "/graph." + index + ".txt");
-
-//        System.out.println(file.getAbsolutePath());
 
         try {
             BufferedReader r = new BufferedReader(new FileReader(file));
