+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/algorithm/external/ExternalAlgorithmTetrad.java
@@ -46,14 +46,11 @@ import java.util.List;
  *
  * @author jdramsey
  */
-public class ExternalAlgorithmTetrad implements ExternalAlgorithm {
+public class ExternalAlgorithmTetrad extends ExternalAlgorithm  {
     static final long serialVersionUID = 23L;
     private final String extDir;
     private String shortDescription = null;
-    private String path;
-    private List<String> usedParameters = new ArrayList<>();
-    private Simulation simulation;
-    private int simIndex = -1;
+
 
     public  ExternalAlgorithmTetrad(String extDir) {
         this.extDir = extDir;
@@ -65,9 +62,8 @@ public class ExternalAlgorithmTetrad implements ExternalAlgorithm {
         this.shortDescription = shortDecription;
     }
 
-    @Override
     /**
-     * Reads in the relevant graph from the file (see above) and returns it.
+     * Reads in the relevant graph from the file and returns it.
      */
     public Graph search(DataModel dataSet, Parameters parameters) {
         int index = -1;
@@ -84,23 +80,17 @@ public class ExternalAlgorithmTetrad implements ExternalAlgorithm {
         }
 
         File file = new File(path, "/results/" + extDir + "/" + (simIndex + 1) + "/graph." + index + ".txt");
-
         System.out.println(file.getAbsolutePath());
-
         Graph graph = GraphUtils.loadGraphTxt(file);
-
         GraphUtils.circleLayout(graph, 225, 200, 150);
-
         return graph;
     }
 
-    @Override
     /**
      * Returns the pattern of the supplied DAG.
      */
     public Graph getComparisonGraph(Graph graph) {
         return new EdgeListGraph(graph);
-//        return SearchGraphUtils.patternForDag(new EdgeListGraph(graph));
     }
 
     public String getDescription() {
@@ -111,38 +101,9 @@ public class ExternalAlgorithmTetrad implements ExternalAlgorithm {
         }
     }
 
-    @Override
-    public List<String> getParameters() {
-        return usedParameters;
-    }
-
-    public int getNumDataModels() {
-        return simulation.getNumDataModels();
-    }
-
-    @Override
     public DataType getDataType() {
         return DataType.Continuous;
     }
-
-    public void setSimulation(Simulation simulation) {
-        this.simulation = simulation;
-    }
-
-    public void setPath(String path) {
-        this.path = path;
-    }
-
-    public void setSimIndex(int simIndex) {
-        this.simIndex = simIndex;
-    }
-
-    @Override
-    public Simulation getSimulation() {
-        return simulation;
-    }
-
-    @Override
     public long getElapsedTime(DataModel dataSet, Parameters parameters) {
         int index = -1;
 
@@ -169,5 +130,4 @@ public class ExternalAlgorithmTetrad implements ExternalAlgorithm {
             throw new IllegalArgumentException();
         }
     }
-
 }
